[physical-executor-metrics] 修复了cpu slot metric里long转double的问题

+++ core/src/main/scala/org/apache/spark/executor/ExecutorSource.scala
@@ -68,7 +68,7 @@ class ExecutorSource(threadPool: ThreadPoolExecutor, executorId: String, reqMemo
   // Gauge for executor physical cpu core slot usage rate.
   metricRegistry.register(MetricRegistry.name("cpu", "cpuSlotUsageRate"), new Gauge[Double] {
     override def getValue: Double =
-      threadPool.getActiveCount() / reqCores
+      threadPool.getActiveCount() / reqCores.toDouble
   })
 
   // Gauge for executor thread pool's actively executing task counts
